+++ Changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Version (NEXT)
+* enhanced operation registry for Braket devices
+* added Braket gate names in qasm3 emission 
 * updated Clay
 
 ## Version 0.21.0