+++ Changelog.md
@@ -4,6 +4,7 @@
 * enhanced operation registry for Braket devices
 * added Braket gate names in qasm3 emission
 * always collect trajectories for hardware simulator
+* use noisy measurments and trajectories in result extraction
 * updated Clay
 
 ## Version 0.21.0